+++ docs/zh-cn/changelog.md
@@ -14,6 +14,7 @@
 
 - **Bug Fix**
   - 🐞 修复 vite环境下iframe沙箱对于插件系统`escapeProperties`不支持的问题。
+  - 🐞 修复 iframe沙箱下jquery通过`$('body').append`创建元素逃逸的问题，[issue 1339](https://github.com/micro-zoe/micro-app/issues/1339)。
 
 
 ### 1.0.0-rc.8